requirements.txt

streamlit 
azure-cognitiveservices-vision-computervision==0.7.0